requirements.txt
yfinance==0.1.59
numpy==1.20.3
requests==2.25.1
yahoo_fin==0.8.8
plotly==4.14.3
cufflinks==0.17.3
pandas==1.2.4
matplotlib==3.4.1
pymongo==3.11.0
fbprophet==0.7.1
environ==1.0
Markdown==3.3.4
pandas_datareader==0.9.0
streamlit==0.82.0
mplfinance==0.12.7a17
FinQuant==0.2.2
Pillow==8.2.0
python-dotenv==0.17.1
